+++ corecomponent/androidcomponent/src/main/res/navigation/navigation.xml
@@ -5,7 +5,7 @@
     app:startDestination="@id/board">
     <fragment
         android:id="@+id/board"
-        android:name="com.techcafe.todone.board.BoardFragment"
+        android:name="com.techcafe.todone.board.ui.board.BoardFragment"
         android:label="Board">
         <action
             android:id="@+id/action_board_fragment_to_auth"
