+++ features/board/src/main/java/com/techcafe/todone/board/di/modules.kt
@@ -1,6 +1,6 @@
 package com.techcafe.todone.board.di
 
-import com.techcafe.todone.board.viewModel.BoardCreateViewModel
+import com.techcafe.todone.board.ui.board.BoardCreateViewModel
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
